move FrontPage to pages directory
