+++ src/App.tsx
@@ -1,8 +1,7 @@
 import React from "react";
-import FrontPage from "./components/FrontPage";
+import FrontPage from "./pages/FrontPage";
 import Login from "./components/Login";
 import { Route, Routes } from "react-router-dom";
-
 import "./App.css";
 import Register from "./components/Register";
