+++ src/pages/FrontPage.tsx
@@ -1,6 +1,6 @@
-import Header from "./Header";
-import Content from "./Content";
-import Footer from "./Footer";
+import Header from "../components/Header";
+import Content from "../components/Content";
+import Footer from "../components/Footer";
 import { useToggle } from "../ts/ToggleHeader";
 
 export const FrontPage = () => {
